backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
